+++ testlib/src/main/resources/java/cleanthat/LiteralsFirstInComparisons.clean.java
@@ -1,0 +1,8 @@
+package eu.solven.cleanthat.engine.java.refactorer.cases.do_not_format_me;
+
+public class LiteralsFirstInComparisonsCases {
+
+	public boolean isHardcoded(String input) {
+		return "hardcoded".equals(input);
+	}
+}