+++ CHANGES.md
@@ -10,6 +10,8 @@ This document is intended for Spotless developers.
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+### Added
+* CleanThat Java Refactorer ([#???](https://github.com/diffplug/spotless/pull/???))
 
 ## [2.34.1] - 2023-02-05
 ### Changes